mix test --stale move to Mix.Compilers.Test and only write manifest on 0 failures
